+++ CHANGELOG.md
@@ -8,5 +8,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Matcher for empty optional
 - Matcher for optional with value
+- Matcher for optional with value matching condition
 
 [Unreleased]: https://github.com/clean-code-rocks/hamcrest-java-optional/commits/main